Сделана корректная обработка ошибок на сервере, также проведена валидация на сервере. Тажке подправлены возвращаемые http коды

+++ src/server/controllers/productsController.js
@@ -1,6 +1,13 @@
 import {Product} from "../models/product.js";
+import {validateProduct} from "../utils/validator.js";
+import mongoose from "mongoose";
 
 async function createProduct(data) {
+    const validationErrors = validateProduct(data);
+    if (validationErrors.length > 0) {
+        throw new Error(validationErrors.join(', '))
+    }
+
     const product = await Product.create(data)
     return product;
 }
@@ -11,11 +18,24 @@ async function getProducts() {
 }
 
 async function deleteProduct(id) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw new Error("неверный Id товара")
+    }
+
     const deleted = await Product.deleteOne({_id: id})
     return deleted
 }
 
 async function updateProduct(id, data) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw new Error("неверный Id товара")
+    }
+
+    const validationErrors = validateProduct(data);
+    if (validationErrors.length > 0) {
+        throw new Error(validationErrors.join(', '))
+    }
+
     const updated = await Product.findByIdAndUpdate(id, data, {
         new: true,
     })
